// types/rate.d.ts
export interface RateObject {
  usd: number;
  jpy: number;
  cny: number;
  chf: number;
  cad: number;
  mxn: number;
  inr: number;
  brl: number;
  rub: number;
  krw: number;
  idr: number;
  try: number;
  sar: number;
  sek: number;
  ngn: number;
  pln: number;
  ars: number;
  nok: number;
  twd: number;
  aed: number;
  thb: number;
  zar: number;
  dkk: number;
  myr: number;
  sgd: number;
  ils: number;
  hkd: number;
  php: number;
  clp: number;
  pkr: number;
  czk: number;
  vnd: number;
  bdt: number;
  huf: number;
  uah: number;
  lkr: number;
  mmk: number;
  bhd: number;
  btc: number;
  xrp: number;
  aud: number;
  gbp: number;
  eth: number;
  eur: number;
  ltc: number;
  nzd: number;
}